scripts/requirements.txt
sentence-transformers>=2.7.0,<3.5
numpy>=1.26
httpx>=0.27
PyYAML>=6.0
tenacity>=8.2
python-slugify>=8.0
